+++ Galaga/Galaga/src/ImageComponent.h
@@ -20,7 +20,7 @@ public:
 	ImageComponent& operator=(ImageComponent&& other) = delete;
 
 	void SetSourceRect(SDL_Rect sourceRect);
-	void SetSourceRect(float x, float y, float w, float h);
+	void SetSourceRect(int x, int y, int w, int h);
 	void SetDestRect(float width, float height);
 
 	void Render() const override;
